requirements.txt
asgiref==3.10.0
attrs==25.4.0
brotli==1.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
fonttools==4.60.1
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
